refactor: make more semantic subservers
