+++ countries/main.py
@@ -1,3 +1,6 @@
+from fastmcp import FastMCP
+from mcp.server.fastmcp.prompts import base
+
 COUNTRIES = [
     "Afghanistan",
     "Akrotiri and Dhekelia",
@@ -264,3 +267,19 @@ COUNTRIES_LOWER = [country.lower() for country in COUNTRIES]
 Defines a list of country names and their lowercase equivalents for easy lookup.
 See also COUNTRIES.  This is a list comprehension of that variable into lowercase.
 """
+
+
+mcp = FastMCP(name="countries")
+
+@mcp.prompt()
+def list_countries_prompt() -> list[base.Message]:
+    """Prompt for listing available countries"""
+    return [
+        base.UserMessage("What are the available countries in the CIA World Factbook? Please list them."),
+        # TODO: Is COUNTRIES a tool? 
+        base.AssistantMessage("Here is the list of countries available in the CIA World Factbook:\n" + "\n".join(COUNTRIES))
+    ]
+
+
+
+PROMPTS = [list_countries_prompt]
