--- server/main.py
@@ -1,4 +1,0 @@
-from fastmcp import FastMCP
-
-# Create an MCP server
-mcp = FastMCP("World Facts Bot")
